requirements.txt
# Core dependencies
tqdm>=4.66.0
beautifulsoup4>=4.12.2
ebooklib>=0.17.1
google-generativeai>=0.3.2
google-api-core>=2.17.0
openai>=1.30.1
tiktoken>=0.6.0
